requirements.txt
aiogram==3.15.0
python-dotenv==1.0.1
openai==1.58.1
sentence-transformers==3.3.1
faiss-cpu==1.9.0
numpy==1.26.4
